+++ app/src/main/java/woowacourse/movie/data/AdRepositories.kt
@@ -1,6 +1,7 @@
 package woowacourse.movie.data
 
 import woowacourse.movie.R
+import woowacourse.movie.model.MovieAndAd
 
 class AdRepositories {
     val ads = List(10) {
