+++ app/src/main/java/woowacourse/movie/main/AdViewHolder.kt
@@ -4,7 +4,7 @@ import android.view.View
 import android.widget.ImageView
 import androidx.recyclerview.widget.RecyclerView
 import woowacourse.movie.R
-import woowacourse.movie.data.MovieAndAd
+import woowacourse.movie.model.MovieAndAd
 
 class AdViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
     private val image: ImageView = itemView.findViewById(R.id.ad_image)
