+++ app/src/main/java/woowacourse/movie/main/MovieViewHolder.kt
@@ -6,7 +6,7 @@ import android.widget.ImageView
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
 import woowacourse.movie.R
-import woowacourse.movie.data.MovieAndAd
+import woowacourse.movie.model.MovieAndAd
 
 class MovieViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
     private val image: ImageView = itemView.findViewById(R.id.image)
